pyproject.toml
[project]
name = "exploration-langchain-studio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "langgraph-cli[inmem]>=0.4.2",
    "python-dotenv>=1.1.1",
]
